pyscript/requirements.txt
absl-py==2.1.0
astunparse==1.6.3
beautifulsoup4==4.12.3
blinker==1.7.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.2.0
cycler==0.12.1
flask==3.0.2
flask-cors==4.0.0
flatbuffers==24.3.25
fonttools==4.50.0
gast==0.5.4
google-pasta==0.2.0
grpcio==1.62.1
h5py==3.10.0
idna==3.6
itsdangerous==2.1.2
jinja2==3.1.3
joblib==1.3.2
keras==3.1.1
kiwisolver==1.4.5
libclang==18.1.1
markdown==3.6
markdown-it-py==3.0.0
markupsafe==2.1.5
matplotlib==3.8.3
mdurl==0.1.2
ml-dtypes==0.3.2
namex==0.0.7
numpy==1.26.4
opt-einsum==3.3.0
optree==0.11.0
packaging==24.0
pandas==2.2.1
pillow==10.2.0
protobuf==4.25.3
pygments==2.17.2
pyparsing==3.1.2
pyqt5==5.15.10
pyqt5-qt5==5.15.2
pyqt5-sip==12.13.0
pyside2==5.13.2
python-dateutil==2.9.0.post0
python-whois==0.9.3
pytz==2024.1
requests==2.31.0
rich==13.7.1
scikit-learn==1.4.1.post1
scipy==1.12.0
seaborn==0.13.2
setuptools==69.2.0
shiboken2==5.13.2
six==1.16.0
soupsieve==2.5
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.1
tensorflow-io-gcs-filesystem==0.36.0
termcolor==2.4.0
threadpoolctl==3.4.0
typing-extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
werkzeug==3.0.1
wheel==0.43.0
wrapt==1.16.0
xgboost==2.0.3
